+++ src/main/java/cart/member/service/MemberService.java
@@ -1,8 +1,8 @@
-package cart.service;
+package cart.member.service;
 
-import cart.dao.MemberDao;
-import cart.dto.MemberDto;
-import cart.entity.MemberEntity;
+import cart.member.dao.MemberDao;
+import cart.member.dto.MemberDto;
+import cart.member.entity.MemberEntity;
 import java.util.List;
 import java.util.stream.Collectors;
 import org.springframework.stereotype.Service;
