+++ src/main/java/cart/product/dto/ProductInsertRequestDto.java
@@ -1,4 +1,4 @@
-package cart.dto;
+package cart.product.dto;
 
 import javax.validation.constraints.NotBlank;
 import javax.validation.constraints.Pattern;
