+++ src/main/java/cart/product/dto/ProductValidator.java
@@ -1,4 +1,4 @@
-package cart.dto;
+package cart.product.dto;
 
 public class ProductValidator {
     public static final String IMAGE_URL_REGEX = "((http[s]?|ftp):\\/\\/)?(?:www\\.)?[-a-zA-Z0-9@:%._\\+~#=가-힣]{1,256}[:|\\.][a-zA-Z0-9()]{1,6}\\b(?:[-a-zA-Z0-9()@:%_\\+,.~#?&\\/=가-힣]*)";
